move session-related details to models/session.py
